refactor(Action): move Action to common
